Move components to correct directory
